# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kanban Auth System API",
    "version": "1.0.0",
    "description": "Sistema Kanban com autenticação e autorização baseada em roles (ADMIN/MEMBER)",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token obtido no login"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "BACKLOG",
              "IN_PROGRESS",
              "REVIEW",
              "DONE"
            ]
          },
          "assignedTo": {
            "type": "integer",
            "nullable": true
          },
          "assignedToName": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "integer"
          },
          "createdByName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "default": "MEMBER"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "CreateTaskRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "assignedTo": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "assignedTo": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "MoveTaskRequest": {
        "type": "object",
        "required": [
          "newStatus"
        ],
        "properties": {
          "newStatus": {
            "type": "string",
            "enum": [
              "BACKLOG",
              "IN_PROGRESS",
              "REVIEW",
              "DONE"
            ]
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Endpoints de autenticação"
    },
    {
      "name": "Tasks",
      "description": "Endpoints de gerenciamento de tasks"
    },
    {
      "name": "Users",
      "description": "Endpoints de gerenciamento de usuários"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registrar novo usuário",
        "description": "Cria uma nova conta de usuário. Role padrão é MEMBER.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Email já cadastrado"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Fazer login",
        "description": "Autentica um usuário e retorna tokens JWT. Rate limit de 5 tentativas a cada 15 minutos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          },
          "429": {
            "description": "Muitas tentativas de login"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Renovar access token",
        "description": "Gera novos access e refresh tokens usando o refresh token atual (token rotation).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens renovados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido ou expirado"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Fazer logout",
        "description": "Revoga o refresh token, invalidando a sessão.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso"
          },
          "401": {
            "description": "Refresh token inválido"
          }
        }
      }
    },
    "/api/tasks": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Criar nova task",
        "description": "Apenas ADMIN pode criar tasks. Tasks sempre começam em BACKLOG.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Apenas administradores podem criar tasks"
          }
        }
      },
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Listar tasks",
        "description": "ADMIN vê todas as tasks. MEMBER vê apenas tasks atribuídas a ele.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Buscar task por ID",
        "description": "ADMIN pode ver qualquer task. MEMBER só pode ver tasks atribuídas a ele.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da task"
          }
        ],
        "responses": {
          "200": {
            "description": "Task encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          },
          "404": {
            "description": "Task não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Atualizar task",
        "description": "Apenas ADMIN pode editar tasks.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da task"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task atualizada com sucesso"
          },
          "403": {
            "description": "Apenas administradores podem editar tasks"
          },
          "404": {
            "description": "Task não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Deletar task",
        "description": "Apenas ADMIN pode deletar tasks.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da task"
          }
        ],
        "responses": {
          "200": {
            "description": "Task deletada com sucesso"
          },
          "403": {
            "description": "Apenas administradores podem deletar tasks"
          },
          "404": {
            "description": "Task não encontrada"
          }
        }
      }
    },
    "/api/tasks/{id}/move": {
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Mover task entre status",
        "description": "Regras de movimento:\n- MEMBER pode mover: BACKLOG → IN_PROGRESS, IN_PROGRESS → REVIEW\n- ADMIN pode mover: REVIEW → DONE, REVIEW → IN_PROGRESS (rejeição)\n- MEMBER só pode mover tasks atribuídas a ele\n- Não é possível pular etapas\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da task"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task movida com sucesso"
          },
          "400": {
            "description": "Transição inválida"
          },
          "403": {
            "description": "Acesso negado ou task não atribuída"
          },
          "404": {
            "description": "Task não encontrada"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Listar usuários",
        "description": "Apenas ADMIN pode listar todos os usuários.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Buscar usuário por ID",
        "description": "ADMIN pode ver qualquer usuário. Usuários podem ver seus próprios dados.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Atualizar usuário",
        "description": "Apenas ADMIN pode atualizar usuários.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso"
          },
          "403": {
            "description": "Apenas administradores podem atualizar usuários"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "409": {
            "description": "Email já cadastrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletar usuário",
        "description": "Apenas ADMIN pode deletar usuários. Não é possível deletar a si mesmo ou o último admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso"
          },
          "400": {
            "description": "Não é possível deletar a si mesmo ou o último admin"
          },
          "403": {
            "description": "Apenas administradores podem deletar usuários"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    }
  }
}